moved auth routes to auth.js file
